Implement widget for displaying toolbar with edit link
Add "ContextualEditPart" that's added to "ContextualEdit" content type,
which is a widget. This widget can be added to the site and will display
some high level meta information and an edit link for the content item
associated to the current request. The widget will only display when the
logged in user has permission to edit the content.